Divide screens
This will divide screens.
