+++ screens/index.js
@@ -1,0 +1,19 @@
+import Details from './Details';
+import ForgetPassword from './ForgotPassword';
+import Gallery from './Gallery';
+import GalleryForm from './GalleryForm';
+import Login from './Login';
+import MainPage from './MainPage';
+import RegistrationForm from './RegistrationForm';
+import Settings from './Settings';
+
+export {
+  Details,
+  ForgetPassword,
+  Gallery,
+  GalleryForm,
+  Login,
+  MainPage,
+  RegistrationForm,
+  Settings,
+};
